rename repository and package to Granular.jl
The motivation is to uphold the general applicability of this package,
and open the possibility of prospective inclusion in the Julia package
repository.  Ice floes are in the code and documentation now referred to
as grains.  As Github still mirrors anders-dc/Granular.jl to
anders-dc/SeaIce.jl, I keep the previous urls for travis-CI, appveyor,
and codecov.